requirements.txt
google-cloud-bigquery[bqstorage,pandas]
pandas~=1.4.1
pytz==2022.1
google-api-python-client~=2.41.0

# For Youtube
google-auth-httplib2

# Sentiment Analysis
google-cloud-language~=2.4.1
# google-cloud-storage
# google-cloud-storage~=2.2.1
google-cloud-secret-manager